+++ semana10/projeto-labeX/labex/src/pages/home/HomePage.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom'
 import Header  from '../../header/Header'
-import { ImgAstronauta, ContainerChamada, ContainerAstronauta,
-    ContainerTitulos, ContainerTextoLogo, ImgTextoLogo, 
-    ContainerTitulosAdm, Button, ImgAstronautaAdm } from './styled'
+
 import Astronauta from '../../img/astronauta-foguete.png'
 import AstronautaAdm from '../../img/astronauta-adm.png'
 import TextoLogo from '../../img/logomarca.png'
-import Back from '../../img/terra-back.jpg'
 
+import { ImgAstronauta, ContainerChamada, ContainerAstronauta,
+    ContainerTitulos, ContainerTextoLogo, ImgTextoLogo, 
+    ContainerTitulosAdm, Button, ImgAstronautaAdm } from './styled' 
 
 const HomePage = () => {
     const history = useHistory()
@@ -27,7 +27,6 @@ const HomePage = () => {
             <ContainerTextoLogo >
                 <ImgTextoLogo src={TextoLogo} />
             </ContainerTextoLogo>
-            
             <ContainerChamada>
                 <ContainerTitulos>
                     <h3> Precisando sair da rotina?</h3>
@@ -47,10 +46,7 @@ const HomePage = () => {
                     <h4> Acesse agora as suas viagens</h4> 
                     <Button onClick={goToAreaAdm}> Área Adm </Button>
                 </ContainerTitulosAdm>
-               
-            </ContainerChamada>
-            
-            
+            </ContainerChamada>   
         </div>
     )
 }
